+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Listener's corner is more responsive on mobile devices.
  - Host request in wwsu.js fails for non-authentication requests. [Issue 62](https://github.com/Lovinity/wwsu/issues/62)
  - Changing announcement type caused display slides to not be removed when they should have been. Now, when announcement type is changed, a websocket remove is sent to the previous type.
+ - If an unscheduled show is on the air and happens to later fall within a scheduled timeframe, its attendance record will be updated accordingly. [Issue 66](https://github.com/Lovinity/wwsu/issues/66)
 
 ### Removed
  - Removed the frozenRemote status errorCheck as we are favoring webRTC audio calls for remote broadcasts instead of shoutcast.